feat: cleanup live queries implementation
