+++ packages/postgraphile/src/server.ts
@@ -3,12 +3,9 @@ import { postgraphileConfig } from '@zeepkist/core/config/postgraphile'
 import { createElysiaTelemetryPlugin } from '@zeepkist/telemetry'
 import { Elysia } from 'elysia'
 import logixlysia from 'logixlysia'
-import { elysiaGrafserv } from './elysiaGrafserv'
-import { createGraphqlHttpHandler } from './graphqlHttp'
-import { createLiveQueryInvalidationPoller } from './liveQueryInvalidationPoller'
-import { createLiveQueryWebSocketHandlers } from './liveQueryWebSocket'
 import { serveGraphiql } from './middleware/serveGraphiql'
 import { createPostGraphileHandler } from './postgraphileOptions'
+import { createPostGraphileRuntime } from './postgraphileRuntime'
 
 export {
 	createPostGraphileHandler,
@@ -41,34 +38,7 @@ const withLogging = postgraphileConfig.requestLogging
 	: new Elysia()
 
 export function buildPostGraphileServer(handler = createPostGraphileHandler()) {
-	const server = handler.createServ(elysiaGrafserv)
-	const graphqlRoute = createGraphqlHttpHandler(server)
-	const poller = createLiveQueryInvalidationPoller(postgraphileConfig.liveQueries)
-	const liveQueryWebSocket = createLiveQueryWebSocketHandlers({
-		schema: Promise.resolve(handler.getSchema()),
-		debounceMs: postgraphileConfig.liveQueries.debounceMs,
-		maxOperations: postgraphileConfig.liveQueries.maxOperations,
-		onActiveChange(active) {
-			if (active) {
-				poller.start(liveQueryWebSocket.invalidate)
-			} else {
-				poller.stop()
-			}
-		},
-		async execute(request, body) {
-			const response = await server.handleGraphQLRequest(request, body)
-			if (!response) {
-				return { errors: [{ message: 'GraphQL response not found' }] }
-			}
-
-			const text = await response.text()
-			try {
-				return JSON.parse(text) as unknown
-			} catch {
-				return { errors: [{ message: text || 'GraphQL response was not JSON' }] }
-			}
-		},
-	})
+	const runtime = createPostGraphileRuntime(handler)
 
 	return new Elysia({
 		aot: true,
@@ -82,18 +52,12 @@ export function buildPostGraphileServer(handler = createPostGraphileHandler()) {
 		.use(withTelemetry)
 		.get('/healthz', () => 'OK')
 		.head('/healthz', () => 'OK')
-		.ws('/', liveQueryWebSocket.handlers)
+		.ws('/', runtime.liveQueryWebSocket)
 		.get('/', ({ request }) => serveGraphiql(request))
 		.get('/graphiql', ({ request }) => redirectToRoot(request))
 		.get('/graphql', ({ request }) => redirectToRoot(request))
-		.all('/ruru-static/*', async ({ request }) => {
-			return (
-				(await server.handleGraphiQLStaticRequest(request)) ??
-				(await serveGraphiql(request)) ??
-				new Response('Not Found', { status: 404 })
-			)
-		})
-		.post('/', graphqlRoute)
-		.options('/', graphqlRoute)
-		.onStop(() => poller.stop())
+		.all('/ruru-static/*', ({ request }) => runtime.serveRuruStatic(request))
+		.post('/', runtime.graphqlRoute)
+		.options('/', runtime.graphqlRoute)
+		.onStop(runtime.stop)
 }
